webpage codebase reformatted
